Move geometry operations to worker
