+++ app/package.json
@@ -25,6 +25,7 @@
     "@turf/dissolve": "^7.1.0",
     "@turf/helpers": "^7.1.0",
     "axios": "^1.7.2",
+    "comlink": "^4.4.2",
     "d3-scale-chromatic": "^3.1.0",
     "idb-keyval": "^6.2.1",
     "lodash": "^4.17.21",
