+++ app/src/app/utils/geometryWorker.types.ts
@@ -1,0 +1,3 @@
+export type GeometryWorkerClass = {
+  parseGeometry: (features: GeoJSON.Feature[]) => {dissolved: GeoJSON.FeatureCollection, centroids: GeoJSON.FeatureCollection}
+}
